requirements.txt
adal==1.2.3
applicationinsights==0.11.9
azure-common==1.1.25
azure-core==1.5.0
azure-graphrbac==0.61.1
azure-identity==1.2.0
azure-mgmt-authorization==0.60.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==9.0.0
azure-mgmt-storage==9.0.0
azureml-automl-core==1.4.0
azureml-core==1.4.0.post2
azureml-dataprep==1.4.6
azureml-dataprep-native==14.1.0
azureml-pipeline==1.4.0
azureml-pipeline-core==1.4.0
azureml-pipeline-steps==1.4.0
azureml-sdk==1.4.0
azureml-telemetry==1.4.0
azureml-train==1.4.0
azureml-train-automl-client==1.4.0
azureml-train-core==1.4.0
azureml-train-restclients-hyperdrive==1.4.0
backports.tempfile==1.0
backports.weakref==1.0.post1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
contextlib2==0.6.0.post1
cryptography==2.9.2
distro==1.5.0
docker==4.2.0
dotnetcore2==2.1.14
entrypoints==0.3
flake8==3.7.9
fusepy==3.0.1
idna==2.9
importlib-metadata==1.6.0
isodate==0.6.0
jeepney==0.4.3
jmespath==0.9.5
joblib==0.14.1
jsonpickle==1.4.1
mccabe==0.6.1
msal==1.2.0
msal-extensions==0.1.3
msrest==0.6.13
msrestazure==0.6.3
ndg-httpsclient==0.5.1
nltk==3.5
numpy==1.18.4
oauthlib==3.1.0
pandas==1.0.3
pathspec==0.8.0
portalocker==1.7.0
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pypiwin32==223
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
regex==2020.5.7
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.15.89
scikit-learn==0.22.2.post1
scipy==1.4.1
SecretStorage==3.1.2
six==1.14.0
sklearn==0.0
tqdm==4.46.0
urllib3==1.25.9
websocket-client==0.57.0
zipp==3.1.0
